+++ src/components/text-copy/index.js
@@ -11,6 +11,10 @@ exports.TextCopy = void 0;
 
 var _slicedToArray2 = _interopRequireDefault(require("@babel/runtime/helpers/slicedToArray"));
 
+var _extends2 = _interopRequireDefault(require("@babel/runtime/helpers/extends"));
+
+var _objectWithoutProperties2 = _interopRequireDefault(require("@babel/runtime/helpers/objectWithoutProperties"));
+
 var _react = _interopRequireWildcard(require("react"));
 
 var _propTypes = _interopRequireDefault(require("prop-types"));
@@ -34,26 +38,27 @@ var fadeIn = (0, _styledComponents.keyframes)(["0%{opacity:0;}100%{opacity:1;}"]
 var Wrapper = _styledComponents.default.div.withConfig({
   displayName: "text-copy__Wrapper",
   componentId: "sc-1ikj7bl-0"
-})(["position:relative;display:flex;align-items:center;justify-content:space-between;border:", " solid ", ";background-color:", ";cursor:pointer;"], (0, _typography.pxToRem)(2), _colorsConfig.default.line1, _colorsConfig.default.background1);
+})(["position:relative;display:flex;align-items:center;justify-content:space-between;border:", " solid ", ";background-color:", ";"], (0, _typography.pxToRem)(2), _colorsConfig.default.line1, _colorsConfig.default.background1);
 
 var StyledText = (0, _styledComponents.default)(function (_ref) {
   var className = _ref.className,
-      children = _ref.children;
-  return _react.default.createElement(_text.Text, {
+      children = _ref.children,
+      others = (0, _objectWithoutProperties2.default)(_ref, ["className", "children"]);
+  return _react.default.createElement(_text.Text, (0, _extends2.default)({
     className: className
-  }, children);
+  }, others), children);
 }).withConfig({
   displayName: "text-copy__StyledText",
   componentId: "sc-1ikj7bl-1"
-})(["padding:", " ", ";", ""], (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(15), function (_ref2) {
+})(["padding:", " ", ";width:100%;", ""], (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(15), function (_ref2) {
   var forceOneLine = _ref2.forceOneLine;
-  return forceOneLine && (0, _styledComponents.css)(["overflow:hidden;white-space:nowrap;text-overflow:ellipsis;"]);
+  return forceOneLine && (0, _styledComponents.css)(["overflow:hidden;white-space:nowrap;"]);
 });
 
 var IconWrapper = _styledComponents.default.div.withConfig({
   displayName: "text-copy__IconWrapper",
   componentId: "sc-1ikj7bl-2"
-})(["display:flex;align-items:center;padding:", ";border-left:", " solid ", ";align-self:stretch;box-sizing:border-box;"], (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(2), _colorsConfig.default.line1);
+})(["display:flex;cursor:pointer;align-items:center;padding:", ";border-left:", " solid ", ";align-self:stretch;box-sizing:border-box;"], (0, _typography.pxToRem)(10), (0, _typography.pxToRem)(2), _colorsConfig.default.line1);
 
 var StyledArrowContainer = (0, _styledComponents.default)(_arrowContainer.ArrowContainer).withConfig({
   displayName: "text-copy__StyledArrowContainer",
@@ -73,6 +78,11 @@ var TextCopy = function TextCopy(_ref3) {
       isMessageShown = _useState2[1];
 
   var textRef = (0, _react.useRef)(null);
+  var selectText = (0, _react.useCallback)(function () {
+    var range = document.createRange();
+    range.selectNode(textRef.current);
+    window.getSelection().addRange(range);
+  });
   var copyText = (0, _react.useCallback)(function () {
     isMessageShown(false);
 
@@ -87,11 +97,7 @@ var TextCopy = function TextCopy(_ref3) {
       range.selectNode(textRef.current);
       window.getSelection().addRange(range);
     } else {
-      var _range = document.createRange();
-
-      _range.selectNode(textRef.current);
-
-      window.getSelection().addRange(_range);
+      selectText();
       document.execCommand('copy');
     }
 
@@ -99,16 +105,16 @@ var TextCopy = function TextCopy(_ref3) {
       return isMessageShown(true);
     }, 1);
   });
-  return _react.default.createElement(_react.default.Fragment, null, _react.default.createElement(Wrapper, {
-    onClick: copyText
-  }, description && _react.default.createElement(_visuallyHidden.VisuallyHidden, null, description), _react.default.createElement(StyledText, {
+  return _react.default.createElement(_react.default.Fragment, null, _react.default.createElement(Wrapper, null, description && _react.default.createElement(_visuallyHidden.VisuallyHidden, null, description), _react.default.createElement(StyledText, {
     weight: "light",
     size: "default",
-    forceOneLine: forceOneLine
+    forceOneLine: forceOneLine,
+    onClick: selectText
   }, _react.default.createElement("span", {
     ref: textRef
   }, children)), _react.default.createElement(IconWrapper, {
-    "aria-hidden": true
+    "aria-hidden": true,
+    onClick: copyText
   }, _react.default.createElement(_copyIcon.CopyIcon, null)), alertMessage && shouldShowMessage && _react.default.createElement(StyledArrowContainer, {
     color: _colorsConfig.default.primary1,
     position: "top",